Experiment metadata storage mechanics
Experiment is not only a git repository for the experiment code - it organizes
the outputs of every single experiment "run" as well. Each run is referred to
as a Snapshot and it has a tangible object representation. Every snapshot is
uniquely identified and allows storing data (outputs, metrics, plots, ...) for
each particular run in a unique location. Every snapshot "belongs" to a commit
in the actual codebase. But while commits are only created when the code is
changed, snapshots are created every time the user runs an experiment. When
that happens, a new Snapshot object is created by the Experiment (and granted
its unique ID and timestamp, as well as hexsha to its corresponding commit)
and appended to its registry, which is then stored in a serialized form as a
JSON file.