Pytorch training backend prototype
Broken down into separate classes with the intent to mix them into
the final thing (for now only PytorchTrainable, with more to come:
PytorchTestable, PytorchEvaluable, probably also a PytorchServer).
Due to the interdependent structure, the separate classes are not
going to work well on their own - only the final PytorchTask makes
any sense.