transaction-based Snapshot update model + missing logging.py
A Snapshot should update its on-disk file as often as it is itself
updated. Therefore, it should not be directly written to; instead
it should offer some method that would indirectly perform the write
and also ensure that the state is immediately reflected on-disk.
This might become rather inefficient in the situation of multiple
sequential writes. It would be more convenient to bundle them into
a single transaction, and perform the serialization after. Another
advantage will arise when all Snapshot writes will be finally made
atomic - when a lock will be acquired on the file. In a transaction
mode, such a lock will only be acquired once and released once,
instead of repeatedly attempting this for multiple writes.
We accomplish this by providing a SnapshotView class, which is a
ContextManager. By calling a specific Snapshot method (for example
"train_storage") a SnapshotView is acquired, which:
* allows writing directly into its respective destination (in the
  above example: snapshot.train_data), and
* ensures that upon exiting the context, state of the Snapshot will
  be serialized.
In the future, this class can be enhanced with locking/releasing
mechanism to provide read&write atomicity.
For example usage, see Logger.store_train [logging.py].